+++ zenodot/src/dd/kms/zenodot/framework/common/ExceptionalFunction.java
@@ -1,0 +1,7 @@
+package dd.kms.zenodot.framework.common;
+
+@FunctionalInterface
+public interface ExceptionalFunction<T, R>
+{
+	R apply(T t) throws Exception;
+}