+++ zenodot/src/dd/kms/zenodot/framework/wrappers/FieldInfo.java
@@ -1,5 +1,6 @@
 package dd.kms.zenodot.framework.wrappers;
 
+import dd.kms.zenodot.api.common.AccessDeniedException;
 import dd.kms.zenodot.api.common.GeneralizedField;
 
 public class FieldInfo extends MemberInfo<GeneralizedField>
@@ -16,17 +17,29 @@ public class FieldInfo extends MemberInfo<GeneralizedField>
 		return member.getType();
 	}
 
-	public Object get(Object instance) throws IllegalAccessException {
+	public Object get(Object instance) throws AccessDeniedException {
 		if (instance == InfoProvider.INDETERMINATE_VALUE) {
 			return InfoProvider.INDETERMINATE_VALUE;
 		}
-		member.setAccessible(true);
-		return member.get(instance);
+		try {
+			member.setAccessible(true);
+			return member.get(instance);
+		} catch (IllegalArgumentException e) {
+			throw e;
+		} catch (Exception e) {
+			throw new AccessDeniedException("Access to field '" + member.getName() + "' has been denied: " + e, e);
+		}
 	}
 
-	public void set(Object instance, Object value) throws IllegalAccessException {
-		member.setAccessible(true);
-		member.set(instance, value);
+	public void set(Object instance, Object value) throws AccessDeniedException {
+		try {
+			member.setAccessible(true);
+			member.set(instance, value);
+		} catch (IllegalArgumentException e) {
+			throw e;
+		} catch (Exception e) {
+			throw new AccessDeniedException("Access to field '" + member.getName() + "' has been denied: " + e, e);
+		}
 	}
 
 	@Override